third_party/blink/web_tests/external/wpt/fetch/api/basic/scheme-blob.sub.any-expected.txt

This is a testharness.js-based test.
[FAIL] Fetching URL.createObjectURL(invalid_type_blob) is OK
  assert_equals: Content-Type is invalid expected "" but got "invalid"
Harness: the test ran to completion.
